swiping right works
