+++ src/main.rs
@@ -13,25 +13,15 @@ fn main() {
 
     
     let mut game_obj: Game = CustomGame::new();
-    game_obj.add_at_position(4, 3, 0);
-    game_obj.add_at_position(4, 3, 3);
-    //game_obj.add_at_position(2, 0, 3);
+    game_obj.add_at_position(2, 0, 0);
+    game_obj.add_at_position(2, 0, 1);
+    game_obj.add_at_position(2, 0, 2);
     //game_obj.add_at_position(4, 0, 3);
     //game_obj.add_at_position(4, 1, 1);
     game_obj.swipe_right();
     let board = game_obj.get_board();
-    print_board(board);
+    game_obj.print_board();
     
-}
-
-//function for testing purposes
-pub fn print_board(board : &Array2D<i64>){
-    for row_iter in board.rows_iter(){
-        for element in row_iter{
-            print!("{} ", element);
-        }
-        println!();
-    }
 }
